+++ tests/test_agent.py
@@ -66,7 +66,8 @@ def test_agent_with_model_server(tmpdir, zipped_moodbot_model,
                                  moodbot_domain, moodbot_metadata):
     fingerprint = 'somehash'
     model_endpoint_config = EndpointConfig.from_dict(
-            {"url": 'http://server.com/model/default_core@latest'}
+            {"url": 'http://server.com/model/default_core@latest',
+             "wait_time_between_pulls": None}
     )
 
     # mock a response that returns a zipped model
@@ -87,3 +88,52 @@ def test_agent_with_model_server(tmpdir, zipped_moodbot_model,
                       for p in agent.policy_ensemble.policies}
     moodbot_policies = set(moodbot_metadata["policy_names"])
     assert agent_policies == moodbot_policies
+
+
+def test_wait_time_between_pulls_from_file(monkeypatch):
+    from future.utils import raise_
+
+    monkeypatch.setattr("rasa_core.agent.start_model_pulling_in_worker",
+                        lambda *args: True)
+    monkeypatch.setattr("rasa_core.agent._update_model_from_server",
+                        lambda *args: raise_(Exception()))
+
+    model_endpoint_config = utils.\
+        read_endpoint_config("data/test_endpoints/model_endpoint.yml", "model")
+
+    rasa_core.agent.\
+        load_from_server(model_server=model_endpoint_config)
+
+
+def test_wait_time_between_pulls_str(monkeypatch):
+    from future.utils import raise_
+
+    monkeypatch.setattr("rasa_core.agent.start_model_pulling_in_worker",
+                        lambda *args: True)
+    monkeypatch.setattr("rasa_core.agent._update_model_from_server",
+                        lambda *args: raise_(Exception()))
+
+    model_endpoint_config = EndpointConfig.from_dict(
+        {"url": 'http://server.com/model/default_core@latest',
+         "wait_time_between_pulls": "10"}
+    )
+
+    rasa_core.agent.\
+        load_from_server(model_server=model_endpoint_config)
+
+
+def test_wait_time_between_pulls_with_not_number(monkeypatch):
+    from future.utils import raise_
+
+    monkeypatch.setattr("rasa_core.agent.start_model_pulling_in_worker",
+                        lambda *args: raise_(Exception()))
+    monkeypatch.setattr("rasa_core.agent._update_model_from_server",
+                        lambda *args: True)
+
+    model_endpoint_config = EndpointConfig.from_dict(
+        {"url": 'http://server.com/model/default_core@latest',
+         "wait_time_between_pulls": "None"}
+    )
+
+    rasa_core.agent.\
+        load_from_server(model_server=model_endpoint_config)